requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
google-auth==2.27.0
requests==2.31.0